Rename the component file to start with capital
